authbot/requirements.txt
discord.py >= 0.9.1
PyMySQL
